Fix: issue with amounts less that 1,00
FossilOrigin-Name: a3fae2b39a09ee1813d4e2ff57d7d42a8d15fa38

+++ src/include/lib/Garradin/Utils.php
@@ -130,7 +130,7 @@ class Utils
         }
 
         $decimals = substr($number, -2);
-        $number = substr($number, 0, -2);
+        $number = (int) substr($number, 0, -2);
 
         return sprintf('%s%s%s', number_format($number, 0, $dec_point, $thousands_sep), $dec_point, $decimals);
     }
